requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv 
python-multipart